Move TermuxTerminalSessionClientBase to com.termux.shared.terminal package
